API

FlightPhysics.ISAState

FlightPhysics.geometric_to_geopotential_altitude

FlightPhysics.geopotential_to_geometric_altitude

FlightPhysics.gravitational_acceleration

FlightPhysics.isa_state

FlightPhysics.ISAState — Type

ISAState{Scalar<:AbstractFloat}

A structure representing Earth's atmospheric state according to the International Standard Atmosphere (ISA) model.

Fields

temperature: Air temperature [K].

pressure: Air pressure [Pa].

density: Air density [kg/m^3].

speed_of_sound: Speed of sound in the air [m/s].

dynamic_viscosity: Dynamic viscosity of the air [Pa s].

Examples

julia> using FlightPhysics; state = FlightPhysics.ISAState{Float64}(288.15, 101325.0, 1.225, 340.294, 1.81206e-5)
FlightPhysics.ISAState{Float64}(288.15, 101325.0, 1.225, 340.294, 1.81206e-5)

FlightPhysics.geometric_to_geopotential_altitude — Method

geometric_to_geopotential_altitude(geometric_altitude::Scalar) where {Scalar<:AbstractFloat}

Convert a geometric altitude to a geopotential altitude.

Arguments

geometric_altitude: The geometric altitude in meters.

Returns

The corresponding geopotential altitude in meters.

Examples

julia> using FlightPhysics; geometric_to_geopotential_altitude(86000.0)
84854.57642868205

FlightPhysics.geopotential_to_geometric_altitude — Method

geopotential_to_geometric_altitude(geopotential_altitude::Scalar) where {Scalar<:AbstractFloat}

Convert a geopotential altitude to a geometric altitude.

Arguments

geopotential_altitude: The geopotential altitude in meters.

Returns

The corresponding geometric altitude in meters.

Examples

julia> using FlightPhysics; geopotential_to_geometric_altitude(86000.0)
87176.77008750994

FlightPhysics.gravitational_acceleration — Method

gravitational_acceleration(altitude::Scalar) where {Scalar<:AbstractFloat}

Calculate the gravitational acceleration at a given altitude.

Arguments

altitude: The altitude above the Earth's surface.

Returns

The gravitational acceleration at the specified altitude.

Examples

julia> using FlightPhysics; gravitational_acceleration(0.0)
9.80665

FlightPhysics.isa_state — Method

isa_state(altitude::Scalar; temperature_offset::Scalar=0.0) -> ISAState

Calculate the atmospheric state at a given altitude according to the International Standard Atmosphere (ISA) model.

Arguments

altitude: The altitude in meters for which the ISA state is to be calculated.

temperature_offset: An optional temperature offset in Kelvin to be added to the calculated temperature. Default is 0.0.

Returns

ISAState: A structure containing the temperature, pressure, and density at the given altitude.

Throws

DomainError: If the altitude is out of the range covered by the ISA model (0 to 86000 meters).

Examples

julia> using FlightPhysics; state = isa_state(10000.0)
FlightPhysics.ISAState{Float64}(223.14999999999998, 26436.233930953953, 0.4127061186074048, 299.46312836140606, 1.468844333610595e-5)